requirements.txt
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-pinecone==0.2.0
pypdf==4.0.0
tiktoken==0.7.0
streamlit==1.38.0
python-dotenv==1.0.1
pinecone-client==5.0.1
sec_edgar_downloader==5.1.0